subdirectory for python test was not working because os.path.join was not specified, all string concats are causing issue in finding path.
typo in toolchain for OpenMPI 2.0.2 with intel

logic for subdirectory was not setup correctly in buildtest.py
Issue was resolved by checking if subdirectory is "config" in config directory, which
implies there is no directories. code_destdir concats code_dir and subdir by default so subdir
needs to be empty string in case there is no subdirectory. This resolves the file path for source files
when compiling from source