adding support for processing modules with FlatNamingScheme. add flag --module-naming-scheme so users can select module scheme from command line. Default is HMNS that is set in setup.sh.
Add prototype functions:
get_appname
get_appversion
get_toolchain_name
get_toolchain_version

Fix all python modules with reference to software and toolchain
variable and use prototype functions to get the appname.

appversion for FNS is slightly different then HMNS because user
will be passing in -s the version+toolchain so we need to strip
this out to get the appversion. In HMNS this is not an issue.