add support for yaml key scheduler, and jobslot. Now yaml configs that have these keys will generate job script based on configuration and will automatically be created without the need of --job-template.
jobscript will use yaml configuration  instead of --job-template when it is specified.
Note that --job-template creates a jobscript with same configuration for all tests
in a particular directory.

Support for creating SLURM and LSF job script. Job scripts are created
even if scheduler is not found when yaml configuration has scheduler options.
As for --job-template it looks if scheduler is present. Need to figure out
event when both scheduler are present or none is present at same time.