adding python example and fixing schema
Any yaml file with a multi line statement must honor spacing, meaning
that it will not work if the spacing for the key is not indented a few
lines under. There are several issues with this current PR with
respect to organization /naming of outputs, but since I have a PR
in for this I am going to leave for the time being.

Signed-off-by: vsoch <vsochat@stanford.edu>