adding start of work for executor bases
buildtest needs to be flexible to allow specification of more than one
kind of executor, whether that be local, slurm, or something else
we are not considering yet (e.g., cloud executor). For this purpose,
I am adding a BuildExecutor class that is able to read in one of
a number of build executor types, which currently include a LocalExecutor
and SlurmExecutor. The variables and customizations should live on
the level of the schemas, read in by the classes, but not need to be
validated by them. I do not have a good understanding of how the executor
interacts with the builder step, and I will need this understanding to
implement the class into the library.

Signed-off-by: vsoch <vsochat@stanford.edu>