adding basic execut steps for a local (base) executor
This pull request adds the basic interaction of an executor with
the builders produced by the BuildTestConfig class. Basically,
we create the BuildExecutors (executor) based on the config_opts
(the users loaded configuration file) and then parse through the builders
and hand each one off to it. In the case of defining an --executor
on the command line, we honor and use that choice. In the case of finding
the executor as a metadata value in the config schema, we also honor that.
In the case of no command option or no default, we either use the first
in the file OR the base class (which is basically a local executor). This
example should be sufficient to show how to set up the same for slurm -
each of steps and dryrun_steps and the subsequent functions would need to
be defined.

Signed-off-by: vsoch <vsochat@stanford.edu>