Refactor BuilderBase to account for changes to shell, shebang key in script schema (#290)
* sync script schema from
https://github.com/buildtesters/schemas/blob/master/script/script-v0.0.1.schema.json
modify implementation on shell detection, environment variable, and the
command to execute the test. We can now support passing shell options on
the command line. The shebang can be customized via 'shebang' key if
it's not found it will use the default shebang based on shell type. If
it's python it will resort to path of 'python' wrapper.
When running in dry mode we were getting output header of table of
results that were not suppose to show up.
Add more log details for BuilderBase

* raise exception BuildTestError when self.shell_name (python, sh, bash)
not found in system. We need to terminate early since we can't run test
without it.

* applying black

Signed-off-by: Shahzeb Siddiqui <shahzebmsiddiqui@gmail.com>

* Add Shell class that returns shell program, options, shell path, and
shebang. This class is intended to get the shell type
based on 'shell' key in Buildspec. The Shell class is not integrated
with BuilderBase class at this point.

* make "bash" as default shell so we can invoke Shell() without any
argument.
Raise exception in init method when path is not found. This raises an
error early as pose to later when invoking Shell.get()

* fixing up shell parser to use shared setter and functions

Signed-off-by: vsoch <vsochat@stanford.edu>

* integrate Shell with BuilderBase class.
Remove env support detection when Buildspec has 'shell: python'. Should
make note that 'env' works for shell scripts and not python scripts.

* applying black

* address fix reported by @vsoch

* Fix two bugs introduced in previous commit.
Need to add self.metadata["shell"] because it was referenced in
test_base.py
The second bug was when a Buildspec doesn't contain 'env' it causes
a bug during environment section, we needed to skip process any keys
since there is no items to process

* add regression test for Shell class, tested this locally, should
expect  100% coverage.

Co-authored-by: vsoch <vsochat@stanford.edu>