first attempt at implementing compiler schema in Buildtest (#300)
* first attempt at implementing 'compiler' schema in Buildtest
Added class CompilerBuilder which is a subclass of BuilderBase that
implements the compiler schema.
Add support for viewing compiler schema via 'buildtest show schema'
Modify global schema since we ran into validation error when specifying
maintainer key. This was due to expected type.
Add compiler schema to buildtest.

* Refactor code by adding lookup dictionary for language file extension
and compiler table. We add exception when any C, C++, Fortran compiler
is not set we raise an error. Replace if/else condition with more
simplified and reduced code structure. Add docstring to method
_build_testcontent
add pgi, cray to compiler schema.

* Add Compiler class GNUCompiler, IntelCompiler, PGICompiler, CrayCompiler
that are subclass of CompilerBuilder which set cc, cxx, fc. We have a
get_compilers method to return the list of compiler wrapper. In
addition, we have get_path method which returns a key/value pair
of compiler wrapper and full path to binary. In future, we may extend
feature to invoke the compiler class by name of module to
create multiple instance of Compilers.
We removed detect_compiler method and refactored language detection
in method detect_lang

* unfinished work, major refactor for CompilerBuilder class to make
class more modular. Right now there are some issues with how test
content is generated.

* move exec_args into compiler property as pose to compiler definition in schema.
2nd attempt at refactoring CompilerBuilder

* add 'name' key in compiler to select the compiler type which
selects the appropriate Compiler class to invoke. Tested this
with intel and gnu compiler and it was successful.

* invoked wrong Compiler class when compiler name is cray

Signed-off-by: Shahzeb Siddiqui <shahzebmsiddiqui@gmail.com>

* use os.path.join instead of using fstring

* remove self.compiler_setup() this method doesnt exist anymore.
Fix os.path.join in building run command