update settings schema, and behavior of BuilderBase and BuildExecutor (#309)
* update settings schema based on https://github.com/buildtesters/schemas/pull/39
Split build and run invocation into two separate calls, rename prepare_run to build_setup
and add few more keys to builder.metadata including name, buildspec, build_id, recipe. This
would be useful since we can use builder.metadata for printing output and this object is passed
to executor class. The builder.metadata['recipe'] is the loaded Buildspec which would be
useful when we need to process Buildspec during execution.

Refactor test examples due to change in BuilderBase and BuildExecutor. Removed
example valid configs right now we only testing combined-example.yml. In future we
can add few more examples but ideally the valid settings should be in sync with
tests conducted in schema repo.

Update example buildspecs in tests/examples/buildspecs to make use of
executor key since we need to specify the name of executor when running job.
In BuildExecutor we check the executor key in Buildspec and see if it matches
one of the executors specified in buildtest setting before we launch test.

Next step would be to move implementation of BuilderBase.run into the LocalExecutor
class to implement the local test executor.

* bug in invoking SSHExecutor when we see slurm key, we should be invoking
SlurmExecutor

* applying black

Signed-off-by: Shahzeb Siddiqui <shahzebmsiddiqui@gmail.com>

* deleting lines that were commented out to clean up some changes i
wanted to keep in case they were to be put back in.