Slurm executor (#329)
* add implementation for dispatch method in SlurmExecutor that is responsible
for submitting job to scheduler. We define methods write_test and check_test_state
in BaseExecutor so it can be reused in subclass.

* update schema with sbatch key in global.schema.json and add key support
in script, compiler schema.

* Add implementation to poll a pending/running job from SlurmExecutor.
The poll is conducted via sacct on a 10sec interval. We are able to
break out of loop once job state is not PENDING or RUNNING. We need
to sort out steps after poll to gather output and error file, and
report job state. to buildtest.

* add gather method to SlurmExecutor class to acquire details from sacct after job completion. We extract ExitCode from output, also we keep track of sacct output internally for reporting purposes

Signed-off-by: Shahzeb Siddiqui <shahzebmsiddiqui@gmail.com>

* Invoke each step in local and slurm executor manually in BuildExecutor run
method. We do this to have control over each step execution. The check method
in SlurmExecutor will check if launcher and sacct are aviable if not we raise
an error before dispatching job.
Add sbatch key support during test generation. The sbatch key is generated at start of script
right after shebang