add definitions for string_or_list, list_of_strings, list_of_ints, int_or_list. (#486)
tags field can be a string or list.
returncode field can be int or list.
Add valid/invalid tests for each field.
Update returncode logic to convert int to list if specified as an integer.
The returncode and tags must be unique items if specified as list. Previously
we could pass duplicate values in list which is incorrect.