change unique testid behavior, add support for --rebuild option (#508)
* We changed unique test id as date timestamp to a uid using uuid.uuid4() the
default is 8 characters for unique id but we have full id available. The --rerun
option is renamed to --rebuild and we set an upperbound to 50.
The report filter field 'returncode' needed check for input to be an integer, previously
there was no check and buildtest exited without proper error message.

We rebuilt all documentation tests, add example for running test in debug mode,
rebuild option, and exceptions to report filter.
We changed path to where generic tests were written however, buildtest buildspec find
wasn't picking this up so we fixed the path name.

We add regression test for --rebuild option and modified test for
BuildTestError. The BuildTestError class was refactored slightly to accept
multiple arguments.

* regenerate script documentation and fix pyflakes issue.
There was error in script examples needed to pass rebuild option
to parse_buildspecs method