add key load_default_buildspecs key in settings schema this configures if buildspecs are loaded (#548)
by default.
Add option --root to buildtest buildspec find command. This allows user to specify a directory or
filepath to load buildspec into cache. If this option is specified we automatically rebuild cache.
If file or directory is invalid, we ignore argument. If file doesn't end in .yml extension we ignore file