change compiler schema where property env, vars, status, pre_build, post_build, pre_run, post_run, run sbatch, bsub, cobalt, batch, BB, DW are moved into compilers section. They can be defined in 'all' section with compilers or some properties can be defined in compiler group (gcc, intel, pgi, etc...) Due to these changes few changes were required in core builder class such as status check. This was implemented in Executor class which required different keys to be looked up.
The generate_script section for ScriptBuilder and CompilerBuilder class now implement entire logic of class,
previously part of the implementation was done in base class which caused some issues.

Added documentation in compiler section.
We also add _EXEC variable to reference executable in generated script, this can be used if user wants
to customize run line.