Update buildspecs_interface.rst
Changed the documentation to reflect the change to the ``buildtest buildspec edit-test`` command. The alias for the buildspec command that allowed you to edit tests based on a supplied test name has been changed from ``buildtest buildspec edit`` to ``buildtest buildspec edit-test``.  This was done in order to better match with the ``buildtest buildspec edit-file``.